feat(server):basic message sent

+++ server/src/entities/sqlite/user_chat_msg.rs
@@ -5,8 +5,8 @@ use sea_orm::entity::prelude::*;
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "user_chat_msg")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub chat_msg_id: i64,
+    #[sea_orm(primary_key)]
+    pub chat_msg_id: i32,
     pub msg_data: String,
     pub sender_id: i64,
     pub session_id: i64,
